@@ -3,64 +3,64 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "npm run dist && source config/env && webpack serve",
-    "build": "npm run dist && webpack --env prod && npm run build:server",
+    "start": "source config/env && webpack serve",
     "start:server": "source config/env && nodemon ./server/index.js",
-    "build:server": "cp -r ./server/* .",
-    "dist": "rm -rf build && mkdir build"
+    "prebuild": "rm -rf build && mkdir build",
+    "build": "webpack --env prod && npm run build:server",
+    "build:server": "cp -r ./server/* ."
   },
   "dependencies": {
     "connect-pg-simple": "^7.0.0",
     "cors": "^2.8.5",
     "crypto-js": "^4.1.1",
-    "epubjs": "^0.3.90",
-    "express": "^4.17.1",
-    "express-rate-limit": "^5.5.1",
+    "epubjs": "^0.3.93",
+    "express": "^4.17.3",
+    "express-rate-limit": "^6.3.0",
     "express-session": "^1.17.2",
     "focus-visible": "^5.2.0",
-    "html2canvas": "^1.3.2",
-    "idb-keyval": "^6.0.3",
+    "html2canvas": "^1.4.1",
+    "idb-keyval": "^6.1.0",
     "normalize.css": "^8.0.1",
-    "pdfjs-dist": "2.10.377",
+    "pdfjs-dist": "2.13.216",
     "pg-promise": "^10.11.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.2",
-    "react-scripts": "4.0.3",
-    "sequelize": "^6.9.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.0",
+    "sequelize": "^6.18.0",
     "uuid": "^8.3.2",
     "validator": "^13.7.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/eslint-parser": "^7.16.3",
-    "@babel/preset-env": "^7.16.0",
-    "@babel/preset-react": "^7.16.0",
-    "babel-loader": "^8.2.3",
-    "copy-webpack-plugin": "^9.1.0",
-    "core-js": "^3.19.1",
-    "css-loader": "^6.5.1",
-    "css-minimizer-webpack-plugin": "^3.1.3",
-    "eslint": "^8.2.0",
-    "eslint-plugin-react": "^7.27.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "@babel/core": "^7.17.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "babel-loader": "^8.2.4",
+    "copy-webpack-plugin": "^10.2.4",
+    "core-js": "^3.21.1",
+    "css-loader": "^6.7.1",
+    "css-minimizer-webpack-plugin": "^3.4.1",
+    "eslint": "^8.12.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.4.0",
     "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.4.4",
+    "mini-css-extract-plugin": "^2.6.0",
     "nodemon": "^2.0.15",
-    "postcss": "^8.3.11",
-    "postcss-import": "^14.0.2",
-    "postcss-loader": "^6.2.0",
-    "postcss-preset-env": "^6.7.0",
-    "sass": "^1.43.4",
-    "sass-loader": "^12.3.0",
-    "stylelint": "^14.1.0",
-    "stylelint-config-standard": "^24.0.0",
-    "stylelint-scss": "^4.0.0",
-    "terser-webpack-plugin": "^5.2.5",
-    "webpack": "^5.64.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.5.0",
-    "workbox-webpack-plugin": "^6.3.0",
+    "postcss": "^8.4.12",
+    "postcss-import": "^14.1.0",
+    "postcss-loader": "^6.2.1",
+    "postcss-preset-env": "^7.4.3",
+    "sass": "^1.49.11",
+    "sass-loader": "^12.6.0",
+    "stylelint": "^14.6.1",
+    "stylelint-config-standard": "^25.0.0",
+    "stylelint-scss": "^4.2.0",
+    "terser-webpack-plugin": "^5.3.1",
+    "webpack": "^5.71.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.8.1",
+    "workbox-webpack-plugin": "^6.5.2",
     "worker-loader": "^3.0.8"
   },
   "browserslist": [
