@@ -1,5 +1,5 @@
 import { StrictMode } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import "focus-visible";
 import "normalize.css";
@@ -7,11 +7,12 @@ import "./styles/index.scss";
 
 import App from "./components/App";
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root"));
+
+root.render(
   <StrictMode>
     <App/>
-  </StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 );
 
 if (process.env.NODE_ENV === "production") {
